fix: optimize manual run
